refactor: keep forge subscription wiring scoped
