+++ internal/tools/forge_tools.go
@@ -492,73 +492,5 @@ func (r *Registry) registerForgeTools() {
 		},
 	})
 
-	// --- Repository subscriptions ---
-
-	r.Register(&Tool{
-		Name:        "forge_repo_follow",
-		Description: "Follow a code forge repository for new releases and/or commits. Polling runs only when forge subscription polling is enabled in config.",
-		Parameters: map[string]any{
-			"type": "object",
-			"properties": map[string]any{
-				"repo": map[string]any{
-					"type":        "string",
-					"description": "Repository name — 'owner/repo' or just 'repo' (uses default owner)",
-				},
-				"account": map[string]any{
-					"type":        "string",
-					"description": "Forge account name (default: primary account)",
-				},
-				"name": map[string]any{
-					"type":        "string",
-					"description": "Optional friendly name for notifications",
-				},
-				"branch": map[string]any{
-					"type":        "string",
-					"description": "Branch/ref to track for commits. Defaults to the repository default branch.",
-				},
-				"track_releases": map[string]any{
-					"type":        "boolean",
-					"description": "Whether to report new releases. Defaults to true.",
-				},
-				"track_commits": map[string]any{
-					"type":        "boolean",
-					"description": "Whether to report new commits on branch/ref. Defaults to true.",
-				},
-			},
-			"required": []string{"repo"},
-		},
-		Handler: func(ctx context.Context, args map[string]any) (string, error) {
-			return r.forgeTools.HandleRepoFollow(ctx, args)
-		},
-	})
-
-	r.Register(&Tool{
-		Name:        "forge_repo_unfollow",
-		Description: "Stop following a code forge repository. Use forge_repo_subscriptions to find the subscription_id.",
-		Parameters: map[string]any{
-			"type": "object",
-			"properties": map[string]any{
-				"subscription_id": map[string]any{
-					"type":        "string",
-					"description": "The subscription identifier returned by forge_repo_follow or forge_repo_subscriptions.",
-				},
-			},
-			"required": []string{"subscription_id"},
-		},
-		Handler: func(ctx context.Context, args map[string]any) (string, error) {
-			return r.forgeTools.HandleRepoUnfollow(ctx, args)
-		},
-	})
-
-	r.Register(&Tool{
-		Name:        "forge_repo_subscriptions",
-		Description: "List code forge repository subscriptions with tracking settings and latest observed release/commit.",
-		Parameters: map[string]any{
-			"type":       "object",
-			"properties": map[string]any{},
-		},
-		Handler: func(ctx context.Context, args map[string]any) (string, error) {
-			return r.forgeTools.HandleRepoSubscriptions(ctx, args)
-		},
-	})
+	r.registerForgeSubscriptionTools()
 }
